requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1

# Rate limiting
slowapi==0.1.9

# RAG pipeline
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# LLM — Primary: Amazon Bedrock; Fallback: Groq
boto3>=1.34.0
botocore>=1.34.0
groq==0.9.0

# TTS
gTTS==2.5.1

# Utilities
python-multipart==0.0.9
httpx==0.27.0
